# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rideshare Blockchain Bridge API
  description: API for managing blockchain-based escrow, payments, referrals, and insurance contributions for a rideshare platform.
  version: 0.2.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/escrow/initiate:
    post:
      summary: Initiate a new ride escrow and optional insurance contribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rideId, riderWallet, driverWallet, amount]
              properties:
                rideId:
                  type: string
                riderWallet:
                  type: string
                driverWallet:
                  type: string
                amount:
                  type: string
                currency:
                  type: string
                  default: ETH
                insuranceContribution:
                  type: string
                referrerId:
                  type: string
      responses:
        '201':
          description: Escrow initiated
...
  /api/escrow/cancel:
    post:
      summary: Cancel a ride and handle refunds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rideId, cancellationType]
              properties:
                rideId:
                  type: string
                cancellationType:
                  type: string
                  enum: [rider, driver, platform]
      responses:
        '200':
          description: Escrow refunded
...